+++ .vscode/launch.json
@@ -11,7 +11,7 @@
             "preLaunchTask": "dotnet: build",
             "program": "${workspaceFolder}/server/bin/Debug/net6.0/server.dll",
             "args": [],
-            "cwd": "${workspaceFolder}",
+            "cwd": "${workspaceFolder}/server/",
             "stopAtEntry": false,
             "serverReadyAction": {
                 "action": "openExternally",
