+++ Olimpo.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.002.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Olimpo", "Olimpo\Olimpo.csproj", "{EA178F41-513A-40CC-84DE-6776593FA094}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "server", "server\server.csproj", "{A2497E57-BDE1-483F-B092-8FDC45C43A4B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +15,10 @@ Global
 		{EA178F41-513A-40CC-84DE-6776593FA094}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EA178F41-513A-40CC-84DE-6776593FA094}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EA178F41-513A-40CC-84DE-6776593FA094}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A2497E57-BDE1-483F-B092-8FDC45C43A4B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A2497E57-BDE1-483F-B092-8FDC45C43A4B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A2497E57-BDE1-483F-B092-8FDC45C43A4B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A2497E57-BDE1-483F-B092-8FDC45C43A4B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
