+++ server/Domain/Channel.cs
