+++ server/Domain/Sensor.cs
