+++ server/Domain/Stack.cs
