+++ server/Sensors/CPU_Temp.cs
