+++ server/Sensors/ISensor.cs
