+++ server/Sensors/ISensor2.cs
