+++ server/Sensors/Ping.cs
