+++ server/SensorsChecker.cs
