+++ server/appsettings.json
@@ -1,9 +1,155 @@
 {
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "AllowedHosts": "*"
+    "Logging": {
+        "LogLevel": {
+            "Default": "Information",
+            "Microsoft.AspNetCore": "Warning"
+        }
+    },
+    "AllowedHosts": "*",
+    "stacks":[
+        {
+            "id":"bf81ddd4-08a6-4b7b-9b0f-fe95fc2da009",
+            "name":"trade",
+            "services":[
+                {
+                    "id":"2766ae2f-ef72-49ac-b38d-6a0346ef3793",
+                    "name":"Self Machine",
+                    "host":"192.168.18.1",
+                    "sensors":[
+                        {
+                            "id":"10d0d0d0-0d0d-0d0d-0d0d-abdb8f279371",
+                            "name":"Ping",
+                            "port": 22,
+                            "type": "ping",
+                            "metric_unit": "ms", 
+                            "check_each": 500,
+                            "num_tries": 5,
+                            "num_min_before_error": 2, 
+                            "timeout": 10000
+                        },
+                        {
+                            "id":"b0d0d0d0-0d0d-0d0d-0d0d-abdb8f279372",
+                            "name":"CPU Usages",
+                            "type":"cpu_usage",
+                            "port": 22,
+                            "num_tries": 5,
+                            "num_min_before_error": 2, 
+                            "timeout": 1000
+                        },
+                        {
+                            "id":"35f305f3-cd00-4b1f-93ed-abdb8f279372",
+                            "name":"CPU Temp",
+                            "type":"cpu_temp"
+                        },
+                        {
+                            "id":"35f305f3-cd00-4b1f-93ed-abdb8f279371",
+                            "name":"Hard Disc Usaded",
+                            "type":"hd_used"
+                        },
+                        {
+                            "id":"35f305f3-cd00-4b1f-93ed-abdb8f279371",
+                            "name":"Hard Disc Temp",
+                            "type":"hd_temp"
+                        }
+                        ,
+                        {
+                            "id":"35f305f3-cd00-4b1f-93ed-abdb8f279371",
+                            "name":"Power On Hours",
+                            "type":"hd_power_on_hours"
+                        }
+                    ]
+                },
+                {
+                    "id":"aa6df4e8-4a07-45f7-bf04-48e6fef46448",
+                    "name":"Shinobi",
+                    "host":"10.10.1.201",
+                    "sensors":[
+                        {
+                            "id":"35f305f3-cd00-4b1f-93ed-abdb8f279371",
+                            "name":"Machine Memory Total",
+                            "type":"SSH_Memory_Total",
+                            "port": 22,
+                            "username": "hudsonventura",
+                            "password": "#MY P@ssw0rd"
+                        },
+                        {
+                            "id":"35f305f3-cd00-4b1f-93ed-abdb8f279371",
+                            "name":"Machine Memory Free",
+                            "type":"SSH_Memory_Free",
+                            "username": "hudsonventura",
+                            "password": "#MY P@ssw0rd"
+                        },
+                        {
+                            "id":"35f305f3-cd00-4b1f-93ed-abdb8f279371",
+                            "name":"Machine Memory Used",
+                            "type":"SSH_Memory_Used",
+                            "username": "hudsonventura",
+                            "password": "#MY P@ssw0rd"
+                        },
+                        {
+                            "id":"10d0d0d0-0d0d-0d0d-0d0d-abdb8f279371",
+                            "name":"Frontend HTTP with SSL",
+                            "port": 8080,
+                            "type": "https",
+                            "SSL_Verification_Check": false,
+                            "check_each": 60000,
+                            "num_tries": 5,
+                            "num_min_before_error": 2, 
+                            "timeout": 1000,
+                            "alerts":{
+                                "type": "exact", //or "upper_better"
+                                "critical": 500,
+                                "warning": 300,
+                                "success": 200
+                            }
+                        },
+                        {
+                            "id":"35f305f3-cd00-4b1f-93ed-abdb8f279371",
+                            "name":"qBittorrent",
+                            "type":"SSH_Docker_Container",
+                            "username": "hudsonventura",
+                            "password": "#MY P@ssw0rd",
+                            "check_each": 60000,
+                            "metric_unit": "MB", 
+                            "alerts":{
+                                "type": "lower_better",
+                                "critical": 950,
+                                "warning": 890,
+                                "success": 500
+                            }
+                        }
+                    ]
+                }
+            ]
+        }
+    ],
+    "ativos":[
+        {
+            "name":"DNS Google",
+            "host": "8.8.8.8",
+            "timeout": "60",
+            "type": "ping"
+        },
+        {
+            "name":"Cloud Flare",
+            "host": "1.1.1.1",
+            "timeout": "20",
+            "type": "ping"
+        },
+        {
+            "name":"Facebook",
+            "host": "facebook.com",
+            "port": 80,
+            "timeout": "50",
+            "type": "tcp"
+        },
+        {
+            "name":"OPUS",
+            "host": "192.168.18.1",
+            "port": 80,
+            "timeout": "10",
+            "type": "tcp"
+        }
+        
+    ]
 }
